# backend/tests/contract/openapi_v1_snapshot.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ScraperLogResponse": {
    "description": "Schema for scraper log response",
    "properties": {
     "completed_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Completed At"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error_message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Message"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "listings_found": {
      "title": "Listings Found",
      "type": "integer"
     },
     "listings_saved": {
      "title": "Listings Saved",
      "type": "integer"
     },
     "scraper_id": {
      "title": "Scraper Id",
      "type": "integer"
     },
     "started_at": {
      "format": "date-time",
      "title": "Started At",
      "type": "string"
     },
     "status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     }
    },
    "required": [
     "id",
     "scraper_id",
     "started_at",
     "completed_at",
     "status",
     "listings_found",
     "listings_saved",
     "error_message",
     "created_at"
    ],
    "title": "ScraperLogResponse",
    "type": "object"
   },
   "ScraperRunRequest": {
    "description": "Schema for manual scraper run",
    "properties": {
     "max_pages": {
      "anyOf": [
       {
        "maximum": 100.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Pages"
     },
     "target_items": {
      "anyOf": [
       {
        "maximum": 1000.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Target Items"
     }
    },
    "title": "ScraperRunRequest",
    "type": "object"
   },
   "ScraperStatsResponse": {
    "description": "Schema for scraper statistics",
    "properties": {
     "active_scrapers": {
      "title": "Active Scrapers",
      "type": "integer"
     },
     "avg_success_rate": {
      "title": "Avg Success Rate",
      "type": "number"
     },
     "inactive_scrapers": {
      "title": "Inactive Scrapers",
      "type": "integer"
     },
     "last_24h_listings": {
      "title": "Last 24H Listings",
      "type": "integer"
     },
     "last_run": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Run"
     },
     "total_listings": {
      "title": "Total Listings",
      "type": "integer"
     },
     "total_scrapers": {
      "title": "Total Scrapers",
      "type": "integer"
     }
    },
    "required": [
     "total_scrapers",
     "active_scrapers",
     "inactive_scrapers",
     "total_listings",
     "last_24h_listings",
     "last_run",
     "avg_success_rate"
    ],
    "title": "ScraperStatsResponse",
    "type": "object"
   },
   "ScraperTargetCreate": {
    "description": "Schema for creating a new scraper target",
    "properties": {
     "domain": {
      "description": "Domain name (e.g., livingethio.com)",
      "maxLength": 255,
      "minLength": 3,
      "title": "Domain",
      "type": "string"
     },
     "enabled": {
      "default": true,
      "description": "Whether scraper is enabled",
      "title": "Enabled",
      "type": "boolean"
     },
     "max_pages": {
      "default": 50,
      "description": "Maximum pages to scrape",
      "maximum": 100.0,
      "minimum": 1.0,
      "title": "Max Pages",
      "type": "integer"
     },
     "schedule": {
      "default": "daily",
      "description": "Scraping schedule (daily, weekly, custom)",
      "title": "Schedule",
      "type": "string"
     },
     "selectors": {
      "description": "CSS selectors configuration",
      "title": "Selectors",
      "type": "object"
     },
     "url_template": {
      "description": "URL template with {page} placeholder",
      "maxLength": 500,
      "minLength": 10,
      "title": "Url Template",
      "type": "string"
     }
    },
    "required": [
     "domain",
     "url_template",
     "selectors"
    ],
    "title": "ScraperTargetCreate",
    "type": "object"
   },
   "ScraperTargetResponse": {
    "description": "Schema for scraper target response",
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "domain": {
      "title": "Domain",
      "type": "string"
     },
     "enabled": {
      "title": "Enabled",
      "type": "boolean"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "last_run": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Run"
     },
     "last_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Status"
     },
     "max_pages": {
      "title": "Max Pages",
      "type": "integer"
     },
     "schedule": {
      "title": "Schedule",
      "type": "string"
     },
     "selectors": {
      "title": "Selectors",
      "type": "object"
     },
     "total_listings": {
      "title": "Total Listings",
      "type": "integer"
     },
     "updated_at": {
      "format": "date-time",
      "title": "Updated At",
      "type": "string"
     },
     "url_template": {
      "title": "Url Template",
      "type": "string"
     }
    },
    "required": [
     "id",
     "domain",
     "url_template",
     "enabled",
     "selectors",
     "schedule",
     "max_pages",
     "last_run",
     "last_status",
     "total_listings",
     "created_at",
     "updated_at"
    ],
    "title": "ScraperTargetResponse",
    "type": "object"
   },
   "ScraperTargetUpdate": {
    "description": "Schema for updating a scraper target",
    "properties": {
     "domain": {
      "anyOf": [
       {
        "maxLength": 255,
        "minLength": 3,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Domain"
     },
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "max_pages": {
      "anyOf": [
       {
        "maximum": 100.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Pages"
     },
     "schedule": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Schedule"
     },
     "selectors": {
      "anyOf": [
       {
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Selectors"
     },
     "url_template": {
      "anyOf": [
       {
        "maxLength": 500,
        "minLength": 10,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url Template"
     }
    },
    "title": "ScraperTargetUpdate",
    "type": "object"
   },
   "ScraperTestRequest": {
    "description": "Schema for testing scraper configuration",
    "properties": {
     "selectors": {
      "title": "Selectors",
      "type": "object"
     },
     "test_page": {
      "default": 1,
      "description": "Page number to test",
      "minimum": 1.0,
      "title": "Test Page",
      "type": "integer"
     },
     "url_template": {
      "title": "Url Template",
      "type": "string"
     }
    },
    "required": [
     "url_template",
     "selectors"
    ],
    "title": "ScraperTestRequest",
    "type": "object"
   },
   "ScraperTestResponse": {
    "description": "Schema for scraper test results",
    "properties": {
     "error_message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Message"
     },
     "items_found": {
      "title": "Items Found",
      "type": "integer"
     },
     "sample_items": {
      "items": {},
      "title": "Sample Items",
      "type": "array"
     },
     "success": {
      "title": "Success",
      "type": "boolean"
     }
    },
    "required": [
     "success",
     "items_found",
     "sample_items",
     "error_message"
    ],
    "title": "ScraperTestResponse",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   },
   "OAuth2PasswordBearer": {
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    },
    "type": "oauth2"
   }
  }
 },
 "info": {
  "contact": {
   "email": "support@valuadis.et",
   "name": "ValuAdis Support",
   "url": "https://valuadis.et/"
  },
  "description": "Ethiopian Property Valuation Platform API with PostGIS spatial support",
  "license": {
   "name": "MIT",
   "url": "https://opensource.org/licenses/MIT"
  },
  "title": "ValuAdis API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/scrapers/": {
   "get": {
    "description": "Get all scraper targets",
    "operationId": "get_all_scrapers_api_v1_scrapers__get",
    "parameters": [
     {
      "in": "query",
      "name": "skip",
      "required": false,
      "schema": {
       "default": 0,
       "title": "Skip",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ScraperTargetResponse"
         },
         "title": "Response Get All Scrapers Api V1 Scrapers  Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get All Scrapers",
    "tags": [
     "Scrapers"
    ]
   },
   "post": {
    "description": "Create new scraper target",
    "operationId": "create_scraper_api_v1_scrapers__post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ScraperTargetCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScraperTargetResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Create Scraper",
    "tags": [
     "Scrapers"
    ]
   }
  },
...
  "/api/v1/scrapers/logs": {
   "get": {
    "description": "Get scraper logs",
    "operationId": "get_scraper_logs_api_v1_scrapers_logs_get",
    "parameters": [
     {
      "in": "query",
      "name": "scraper_id",
      "required": false,
      "schema": {
       "title": "Scraper Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "skip",
      "required": false,
      "schema": {
       "default": 0,
       "title": "Skip",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ScraperLogResponse"
         },
         "title": "Response Get Scraper Logs Api V1 Scrapers Logs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Scraper Logs",
    "tags": [
     "Scrapers"
    ]
   }
  },
  "/api/v1/scrapers/stats": {
   "get": {
    "description": "Get scraper statistics",
    "operationId": "get_scraper_stats_api_v1_scrapers_stats_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScraperStatsResponse"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Scraper Stats",
    "tags": [
     "Scrapers"
    ]
   }
  },
  "/api/v1/scrapers/{scraper_id}": {
   "delete": {
    "description": "Delete scraper target",
    "operationId": "delete_scraper_api_v1_scrapers__scraper_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "scraper_id",
      "required": true,
      "schema": {
       "title": "Scraper Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Delete Scraper",
    "tags": [
     "Scrapers"
    ]
   },
   "get": {
    "description": "Get scraper by ID",
    "operationId": "get_scraper_api_v1_scrapers__scraper_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "scraper_id",
      "required": true,
      "schema": {
       "title": "Scraper Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScraperTargetResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Scraper",
    "tags": [
     "Scrapers"
    ]
   },
   "put": {
    "description": "Update scraper target",
    "operationId": "update_scraper_api_v1_scrapers__scraper_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "scraper_id",
      "required": true,
      "schema": {
       "title": "Scraper Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ScraperTargetUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScraperTargetResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Update Scraper",
    "tags": [
     "Scrapers"
    ]
   }
  },
  "/api/v1/scrapers/{scraper_id}/run": {
   "post": {
    "description": "Manually trigger scraper run",
    "operationId": "run_scraper_api_v1_scrapers__scraper_id__run_post",
    "parameters": [
     {
      "in": "path",
      "name": "scraper_id",
      "required": true,
      "schema": {
       "title": "Scraper Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "allOf": [
         {
          "$ref": "#/components/schemas/ScraperRunRequest"
         }
        ],
        "title": "Run Data"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Run Scraper",
    "tags": [
     "Scrapers"
    ]
   }
  },
  "/api/v1/scrapers/{scraper_id}/test": {
   "post": {
    "description": "Test scraper configuration",
    "operationId": "test_scraper_api_v1_scrapers__scraper_id__test_post",
    "parameters": [
     {
      "in": "path",
      "name": "scraper_id",
      "required": true,
      "schema": {
       "title": "Scraper Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "allOf": [
         {
          "$ref": "#/components/schemas/ScraperTestRequest"
         }
        ],
        "title": "Test Data"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScraperTestResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Test Scraper",
    "tags": [
     "Scrapers"
    ]
   }
  },
  "/api/v1/scrapers/{scraper_id}/toggle": {
   "patch": {
    "description": "Toggle scraper enabled status",
    "operationId": "toggle_scraper_api_v1_scrapers__scraper_id__toggle_patch",
    "parameters": [
     {
      "in": "path",
      "name": "scraper_id",
      "required": true,
      "schema": {
       "title": "Scraper Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScraperTargetResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Toggle Scraper",
    "tags": [
     "Scrapers"
    ]
   }
  },
...
 },
...
}